feat: shim to ns-package-graph working locally
It's a trick really, I proxyquire all of ns-package-graph but snub out the nsolid-cli for my own custom function.

When the `.run` function is run, the first `info` call is pretty basic and just wants to move onto the actual `package_info` call. The package info is resolved using snyk-resolve-deps (which is similar to `npm ls --json` except gives me greater control over the deps).

I then massaged the data into the format that ns-package-graph wants (based on https://git.io/) and then it spits out the data.